requirements.txt
asgiref==3.8.1
Django~=5.1.1
django-cors-headers==4.4.0
django-redis==5.4.0
djangorestframework==3.15.2
mysqlclient==2.2.4
numpy==2.1.2
redis==5.1.1
setuptools==75.1.0
sqlparse==0.5.1
uuid==1.30
wheel==0.44.0
PyJWT~=2.9.0
cryptography~=43.0.1
pillow~=10.4.0
python-dateutil~=2.9.0.post0